finished

+++ src/app/tasks/business-rules/delete-task.handler.ts
@@ -1,8 +1,11 @@
 import { TaskRepository } from './../repositories/task.repository';
 import { Injectable } from '@angular/core';
 import { ITask } from '../models/itask';
-import { DialogService } from '../shared/Services/dialog.service';
 import { ConfirmationService } from 'src/app/core/services/confirmation.service';
+import { MatDialog } from '@angular/material/dialog';
+import { ReasonPopupComponent } from './../../core/services/reason-popup/reason-popup.component';
+import { ReasonPopupService } from 'src/app/core/services/reason-popup.service';
+
 
 @Injectable({
   providedIn: 'root',
@@ -11,7 +14,8 @@ export class DeleteTaskHandler {
   constructor(
     private repository: TaskRepository,
     private confirmation: ConfirmationService,
-    private dialog: DialogService
+    private message: MatDialog,
+    private reasonService: ReasonPopupService
   ) {}
 
   async execute(task: ITask): Promise<boolean> {
@@ -20,12 +24,27 @@ export class DeleteTaskHandler {
         .confirmDelete(`${task.id} - ${task.title}`)
         .subscribe(async (confirmed) => {
           if (confirmed) {
-            await this.repository.delete(task.id);
-            this.dialog
-              .openConfirmDialog('Are you really want to delete/update/create?')
-              .afterClosed();
-          } 
-          resolve(confirmed) ;
+             this.reasonService
+             .openReasonDialog('Please describe the reason')
+             .afterClosed()
+             .subscribe(async (res) => {
+               if (res) {
+
+
+
+                 await this.repository.delete(task.id);
+                 /*  this.confirmation
+                   .confirmDelete('Are you really want to delete/update/create?') */
+
+                 resolve(confirmed);
+
+               }
+             })
+
+
+
+
+          }
         });
     });
   }
